delete ek60 io and parsers in favor for ek80 versions
