+++ echopype/convert/utils/ek_raw_io.py
@@ -1,31 +1,53 @@
-"""
-Code originally developed for pyEcholab
-(https://github.com/CI-CMG/pyEcholab) by NOAA AFSC.
-Contains class ``RawSimradFile`` for low-level file access to read EK60 data files.
-Called by class ConvertEK60 in ``echopype/convert/ek60.py``.
+# coding=utf-8
+
+#     National Oceanic and Atmospheric Administration (NOAA)
+#     Alaskan Fisheries Science Center (AFSC)
+#     Resource Assessment and Conservation Engineering (RACE)
+#     Midwater Assessment and Conservation Engineering (MACE)
+
+#  THIS SOFTWARE AND ITS DOCUMENTATION ARE CONSIDERED TO BE IN THE PUBLIC DOMAIN
+#  AND THUS ARE AVAILABLE FOR UNRESTRICTED PUBLIC USE. THEY ARE FURNISHED "AS IS."
+#  THE AUTHORS, THE UNITED STATES GOVERNMENT, ITS INSTRUMENTALITIES, OFFICERS,
+#  EMPLOYEES, AND AGENTS MAKE NO WARRANTY, EXPRESS OR IMPLIED, AS TO THE USEFULNESS
+#  OF THE SOFTWARE AND DOCUMENTATION FOR ANY PURPOSE. THEY ASSUME NO RESPONSIBILITY
+#  (1) FOR THE USE OF THE SOFTWARE AND DOCUMENTATION; OR (2) TO PROVIDE TECHNICAL
+#  SUPPORT TO USERS.
+
+'''
+.. module:: echolab2.instruments.util.simrad_raw_file
+
+    :synopsis:  A low-level interface for SIMRAD ".raw" formatted files
+
+    Provides the RawSimradFile class, a low-level object for
+        interacting with SIMRAD RAW formated datafiles.
 
 | Developed by:  Zac Berkowitz <zac.berkowitz@gmail.com> under contract for
 | National Oceanic and Atmospheric Administration (NOAA)
 | Alaska Fisheries Science Center (AFSC)
 | Midwater Assesment and Conservation Engineering Group (MACE)
+|
+| Author:
+|       Zac Berkowitz <zac.berkowitz@gmail.com>
+| Maintained by:
+|       Rick Towler   <rick.towler@noaa.gov>
 
-TODO: fix docstring
-"""
+$Id$
+'''
 
 from io import BufferedReader, FileIO, SEEK_SET, SEEK_CUR, SEEK_END
 import struct
 import logging
-from . import ek60_raw_parsers as parsers
+import echopype.convert.utils.ek_raw_parsers as parsers
 
-__all__ = ['RawSimradFile', 'SimradEOF']
+__all__ = ['RawSimradFile']
 
 log = logging.getLogger(__name__)
-
 
 class SimradEOF(Exception):
 
     def __init__(self, message='EOF Reached!'):
         self.message = message
+
 
     def __str__(self):
         return self.message
@@ -40,9 +62,10 @@ class DatagramSizeError(Exception):
         self.file_pos_bytes = file_pos[0]
         self.file_pos_dgrams = file_pos[1]
 
+
     def __str__(self):
         errstr = self.message + '%s != %s @ (%s, %s)' % (self.expected_size, self.retrieved_size,
-                                                         self.file_pos_bytes, self.file_pos_dgrams)
+            self.file_pos_bytes, self.file_pos_dgrams)
         return errstr
 
 
@@ -55,6 +78,7 @@ class DatagramReadError(Exception):
         self.file_pos_bytes = file_pos[0]
         self.file_pos_dgrams = file_pos[1]
 
+
     def __str__(self):
         errstr = [self.message]
         if self.expected_size is not None:
@@ -66,20 +90,26 @@ class DatagramReadError(Exception):
 
 
 class RawSimradFile(BufferedReader):
-    """
+    '''
     A low-level extension of the built in python file object allowing the reading/writing
-    of SIMRAD RAW files on datagram by datagram basis (instead of at the byte level)
-    """
+    of SIMRAD RAW files on datagram by datagram basis (instead of at the byte level.)
 
+    Calls to the read method return parse datagrams as dicts.
+    '''
     #: Dict object with datagram header/python class key/value pairs
     DGRAM_TYPE_KEY = {'RAW': parsers.SimradRawParser(),
                       'CON': parsers.SimradConfigParser(),
                       'TAG': parsers.SimradAnnotationParser(),
                       'NME': parsers.SimradNMEAParser(),
                       'BOT': parsers.SimradBottomParser(),
-                      'DEP': parsers.SimradDepthParser()}
+                      'DEP': parsers.SimradDepthParser(),
+                      'XML': parsers.SimradXMLParser(),
+                      'FIL': parsers.SimradFILParser(),
+                      'MRU': parsers.SimradMRUParser(),
+                      }
 
-    def __init__(self, name, mode='rb', closefd=True, return_raw=False, buffer_size=1024 * 1024):
+
+    def __init__(self, name, mode='rb', closefd=True, return_raw=False, buffer_size=1024*1024):
 
         #  9-28-18 RHT: Changed RawSimradFile to implement BufferedReader instead of
         #  io.FileIO to increase performance.
@@ -93,6 +123,7 @@ class RawSimradFile(BufferedReader):
         self._total_dgram_count = None
         self._return_raw = return_raw
 
+
     def _seek_bytes(self, bytes_, whence=0):
         '''
         :param bytes_: byte offset
@@ -105,12 +136,14 @@ class RawSimradFile(BufferedReader):
 
         BufferedReader.seek(self, bytes_, whence)
 
+
     def _tell_bytes(self):
         '''
         Returns the file pointer position in bytes.
         '''
 
         return BufferedReader.tell(self)
+
 
     def _read_dgram_size(self):
         '''
@@ -121,9 +154,10 @@ class RawSimradFile(BufferedReader):
         if len(buf) != 4:
             self._seek_bytes(-len(buf), SEEK_CUR)
             raise DatagramReadError('Short read while getting dgram size', (4, len(buf)),
-                                    file_pos=(self._tell_bytes(), self.tell()))
+                file_pos=(self._tell_bytes(), self.tell()))
         else:
-            return struct.unpack('=l', buf)[0]  # This return value is an int object.
+            return struct.unpack('=l', buf)[0] #This return value is an int object.
+
 
     def _bytes_remaining(self):
         old_pos = self._tell_bytes()
@@ -134,29 +168,35 @@ class RawSimradFile(BufferedReader):
 
         return offset
 
+
     def _read_timestamp(self):
-        """Attempts to read the datagram timestamp.
-        """
+        '''
+        Attempts to read the datagram timestamp.
+        '''
 
         buf = self._read_bytes(8)
         if len(buf) != 8:
             self._seek_bytes(-len(buf), SEEK_CUR)
             raise DatagramReadError('Short read while getting timestamp',
-                                    (8, len(buf)), file_pos=(self._tell_bytes(), self.tell()))
+                (8, len(buf)), file_pos=(self._tell_bytes(), self.tell()))
 
         else:
             lowDateField, highDateField = struct.unpack('=2L', buf)
-            return lowDateField, highDateField  # integers
+            #  11/26/19 - RHT - modified to return the raw bytes
+            return lowDateField, highDateField, buf
+
 
     def _read_dgram_header(self):
-        """
+        '''
         :returns: dgram_size, dgram_type, (low_date, high_date)
 
-        Attempts to read the datagram header consisting of
-        long dgram_size
-        char[4] type
-        (long lowDateField, long highDateField)
-        """
+        Attempts to read the datagram header consisting of:
+
+            long        dgram_size
+            char[4]     type
+            long        lowDateField
+            long        highDateField
+        '''
 
         try:
             dgram_size = self._read_dgram_size()
@@ -166,6 +206,7 @@ class RawSimradFile(BufferedReader):
             else:
                 raise
 
+        #  get the datagram type
         buf = self._read_bytes(4)
 
         if len(buf) != 4:
@@ -174,15 +215,26 @@ class RawSimradFile(BufferedReader):
             else:
                 self._seek_bytes(-len(buf), SEEK_CUR)
                 raise DatagramReadError('Short read while getting dgram type', (4, len(buf)),
-                                        file_pos=(self._tell_bytes(), self.tell()))
+                    file_pos=(self._tell_bytes(), self.tell()))
         else:
             dgram_type = buf
-
         dgram_type = dgram_type.decode()
 
-        lowDateField, highDateField = self._read_timestamp()
+        #  11/26/19 - RHT
+        #  As part of the rewrite of read to remove the reverse seeking,
+        #  store the raw header bytes so we can prepend them to the raw
+        #  data bytes and pass it all to the parser.
+        raw_bytes = buf
 
-        return dict(size=dgram_size, type=dgram_type, low_date=lowDateField, high_date=highDateField)
+        #  read the timestamp - this method was also modified to return
+        #  the raw bytes
+        lowDateField, highDateField, buf = self._read_timestamp()
+
+        #  add the timestamp bytes to the raw_bytes string
+        raw_bytes += buf
+
+        return dict(size=dgram_size, type=dgram_type, low_date=lowDateField, high_date=highDateField, raw_bytes=raw_bytes)
+
 
     def _read_bytes(self, k):
         '''
@@ -191,6 +243,7 @@ class RawSimradFile(BufferedReader):
 
         return BufferedReader.read(self, k)
 
+
     def _read_next_dgram(self):
         '''
         Attempts to read the next datagram from the file.
@@ -198,90 +251,116 @@ class RawSimradFile(BufferedReader):
         Returns the datagram as a raw string
         '''
 
+        #  11/26/19 - RHT - Modified this method so it doesn't "peek"
+        #  at the next datagram before reading which was inefficient.
+        #  To minimize changes to the code, methods to read the header
+        #  and timestamp were modified to return the raw bytes which
+        #  allows us to pass them onto the parser without having to
+        #  rewind and read again as was previously done.
+
+        #  store our current location in the file
         old_file_pos = self._tell_bytes()
 
-        #  We've come across one instance where the timestamp is (0L, 0L)
-        #  So... now we check every single datagram for this and skip if needed
-
+        #  try to read the header of the next datagram
         try:
-            # _, dgram_type, (low_date, high_date) = self.peek()[:3]
-            header = self.peek()
-
+            header = self._read_dgram_header()
         except DatagramReadError as e:
             e.message = 'Short read while getting raw file datagram header'
             raise e
 
+        #  check for invalid time data
         if (header['low_date'], header['high_date']) == (0, 0):
-            log.warning('Skipping %s datagram w/ timestamp of (0, 0) at %sL:%d', header['type'],
-                        str(self._tell_bytes()), self.tell())
+            log.warning('Skipping %s datagram w/ timestamp of (0, 0) at %sL:%d', header['type'], str(self._tell_bytes()), self.tell())
             self.skip()
             return self._read_next_dgram()
 
-        # _ = self._read_dgram_size()
-        self._seek_bytes(4, SEEK_CUR)
-
+        #  basic sanity check on size
         if header['size'] < 16:
+            #  size can't be smaller than the header size
             log.warning('Invalid datagram header: size: %d, type: %s, nt_date: %s.  dgram_size < 16',
-                        header['size'], header['type'], str((header['low_date'], header['high_date'])))
+                header['size'], header['type'], str((header['low_date'], header['high_date'])))
 
+            #  see if we can find the next datagram
             self._find_next_datagram()
+
+            #  and then return that
             return self._read_next_dgram()
 
-        raw_dgram = self._read_bytes(header['size'])
+        #  get the raw bytes from the header
+        raw_dgram = header['raw_bytes']
+
+        #  and append the rest of the datagram - we subtract 12
+        #  since we have already read 12 bytes: 4 for type and
+        #  8 for time.
+        raw_dgram += self._read_bytes(header['size'] - 12)
+
+        #  determine the size of the payload in bytes
         bytes_read = len(raw_dgram)
 
+        #  and make sure it checks out
         if bytes_read < header['size']:
-            # self._seek_bytes(old_file_pos, SEEK_SET)
-            # raise DatagramReadError('Short read while getting dgram data',
-            #                        (header['size'], len(raw_dgram)), (old_file_pos, self.tell()))
             log.warning('Datagram %d (@%d) shorter than expected length:  %d < %d', self.tell(),
                         old_file_pos, bytes_read, header['size'])
             self._find_next_datagram()
             return self._read_next_dgram()
 
+        #  now read the trailing size value
         try:
             dgram_size_check = self._read_dgram_size()
-
         except DatagramReadError as e:
             self._seek_bytes(old_file_pos, SEEK_SET)
             e.message = 'Short read while getting trailing raw file datagram size for check'
             raise e
 
+        #  make sure they match
         if header['size'] != dgram_size_check:
             # self._seek_bytes(old_file_pos, SEEK_SET)
             log.warning('Datagram failed size check:  %d != %d @ (%d, %d)',
-                        header['size'], dgram_size_check, self._tell_bytes(), self.tell())
+                header['size'], dgram_size_check, self._tell_bytes(), self.tell())
             log.warning('Skipping to next datagram...')
             self._find_next_datagram()
 
             return self._read_next_dgram()
 
+        #  add the header (16 bytes) and repeated size (4 bytes) to the payload
+        #  bytes to get the total bytes read for this datagram.
+        bytes_read = bytes_read + 20
+
         if self._return_raw:
             self._current_dgram_offset += 1
             return raw_dgram
         else:
-            nice_dgram = self._convert_raw_datagram(raw_dgram)
+            nice_dgram = self._convert_raw_datagram(raw_dgram, bytes_read)
             self._current_dgram_offset += 1
             return nice_dgram
 
-    def _convert_raw_datagram(self, raw_datagram_string):
+
+    def _convert_raw_datagram(self, raw_datagram_string, bytes_read):
         '''
         :param raw_datagram_string: bytestring containing datagram (first 4
             bytes indicate datagram type, such as 'RAW0')
         :type raw_datagram_string: str
 
+        :param bytes_read: integer specifying the datagram size, including header
+            in bytes,
+        :type bytes_read: int
+
         Returns a formated datagram object using the data in raw_datagram_string
         '''
+
+        #  11/26/19 - RHT - Modified this method to pass through the number of
+        #  bytes read so we can bubble that up to the user.
 
         dgram_type = raw_datagram_string[:3].decode()
         try:
             parser = self.DGRAM_TYPE_KEY[dgram_type]
         except KeyError:
-            # raise KeyError('Unknown datagram type %s, valid types: %s' % (str(dgram_type), str(self.DGRAM_TYPE_KEY.keys())))
+            #raise KeyError('Unknown datagram type %s, valid types: %s' % (str(dgram_type), str(self.DGRAM_TYPE_KEY.keys())))
             return raw_datagram_string
 
-        nice_dgram = parser.from_string(raw_datagram_string)
+        nice_dgram = parser.from_string(raw_datagram_string, bytes_read)
         return nice_dgram
+
 
     def _set_total_dgram_count(self):
         '''
@@ -291,10 +370,9 @@ class RawSimradFile(BufferedReader):
         :raises: ValueError if self._total_dgram_count is not None (it has been set before)
         '''
         if self._total_dgram_count is not None:
-            raise ValueError(
-                'self._total_dgram_count has already been set.  Call .reset() first if you really want to recount')
+            raise ValueError('self._total_dgram_count has already been set.  Call .reset() first if you really want to recount')
 
-        # Save current position for later
+        #Save current position for later
         old_file_pos = self._tell_bytes()
         old_dgram_offset = self.tell()
 
@@ -308,24 +386,27 @@ class RawSimradFile(BufferedReader):
                 self._total_dgram_count = self.tell()
                 break
 
-        # Return to where we started
+        #Return to where we started
         self._seek_bytes(old_file_pos, SEEK_SET)
         self._current_dgram_offset = old_dgram_offset
+
 
     def at_eof(self):
         old_pos = self._tell_bytes()
         self._seek_bytes(0, SEEK_END)
         eof_pos = self._tell_bytes()
 
-        # Check to see if we're at the end of file and raise EOF
+        #Check to see if we're at the end of file and raise EOF
         if old_pos == eof_pos:
             return True
 
-        # Othereise, go back to where we were and re-raise the original exception
+        #Othereise, go back to where we were and re-raise the original
+        #exception
         else:
             offset = old_pos - eof_pos
             self._seek_bytes(offset, SEEK_END)
             return False
+
 
     def read(self, k):
         '''
@@ -363,6 +444,7 @@ class RawSimradFile(BufferedReader):
         elif k < 0:
             return self.readall()
 
+
     def readall(self):
         '''
         Reads the entire file from the beginning and returns a list of datagrams.
@@ -376,21 +458,24 @@ class RawSimradFile(BufferedReader):
 
         return dgram_list
 
+
     def _find_next_datagram(self):
         old_file_pos = self._tell_bytes()
         log.warning('Attempting to find next valid datagram...')
 
         while self.peek()['type'][:3] not in list(self.DGRAM_TYPE_KEY.keys()):
-            self._seek_bytes(1, 1)
+            self._seek_bytes(1,1)
 
         log.warning('Found next datagram:  %s', self.peek())
         log.warning('Skipped ahead %d bytes', self._tell_bytes() - old_file_pos)
+
 
     def tell(self):
         '''
         Returns the current file pointer offset by datagram number
         '''
         return self._current_dgram_offset
+
 
     def peek(self):
         '''
@@ -401,13 +486,18 @@ class RawSimradFile(BufferedReader):
         '''
 
         dgram_header = self._read_dgram_header()
-        if dgram_header['type'].startswith('RAW'):
+        if dgram_header['type'].startswith('RAW0'):
             dgram_header['channel'] = struct.unpack('h', self._read_bytes(2))[0]
             self._seek_bytes(-18, SEEK_CUR)
+        elif dgram_header['type'].startswith('RAW3'):
+            chan_id = struct.unpack('128s', self._read_bytes(128))
+            dgram_header['channel_id'] = chan_id.strip('\x00')
+            self._seek_bytes(-(16 + 128), SEEK_CUR)
         else:
             self._seek_bytes(-16, SEEK_CUR)
 
         return dgram_header
+
 
     def __next__(self):
         '''
@@ -415,6 +505,7 @@ class RawSimradFile(BufferedReader):
         '''
 
         return self.read(1)
+
 
     def prev(self):
         '''
@@ -425,6 +516,7 @@ class RawSimradFile(BufferedReader):
         raw_dgram = self.read(1)
         self.skip_back()
         return raw_dgram
+
 
     def skip(self):
         '''
@@ -437,22 +529,23 @@ class RawSimradFile(BufferedReader):
 
         if header['size'] < 16:
             log.warning('Invalid datagram header: size: %d, type: %s, nt_date: %s.  dgram_size < 16',
-                        header['size'], header['type'], str((header['low_date'], header['high_date'])))
+                header['size'], header['type'], str((header['low_date'], header['high_date'])))
 
             self._find_next_datagram()
 
         else:
-            self._seek_bytes(header['size'] + 4, SEEK_CUR)
+            self._seek_bytes(header['size']+4, SEEK_CUR)
             dgram_size_check = self._read_dgram_size()
 
             if header['size'] != dgram_size_check:
                 log.warning('Datagram failed size check:  %d != %d @ (%d, %d)',
-                            header['size'], dgram_size_check, self._tell_bytes(), self.tell())
+                    header['size'], dgram_size_check, self._tell_bytes(), self.tell())
                 log.warning('Skipping to next datagram... (in skip)')
 
                 self._find_next_datagram()
 
         self._current_dgram_offset += 1
+
 
     def skip_back(self):
         '''
@@ -468,7 +561,7 @@ class RawSimradFile(BufferedReader):
 
         dgram_size_check = self._read_dgram_size()
 
-        # Seek to the beginning of the datagram and read as normal
+        #Seek to the beginning of the datagram and read as normal
         try:
             self._seek_bytes(-(8 + dgram_size_check), SEEK_CUR)
         except IOError:
@@ -490,6 +583,7 @@ class RawSimradFile(BufferedReader):
 
         self._current_dgram_offset -= 1
 
+
     def iter_dgrams(self):
         '''
         Iterates through the file, repeatedly calling self.next() until
@@ -508,18 +602,21 @@ class RawSimradFile(BufferedReader):
 
             yield new_dgram
 
-    # Unsupported members
+
+    #Unsupported members
     def readline(self):
         '''
         aliased to self.next()
         '''
         return next(self)
 
+
     def readlines(self):
         '''
         aliased to self.read(-1)
         '''
         return self.read(-1)
+
 
     def seek(self, offset, whence):
         '''
@@ -537,10 +634,10 @@ class RawSimradFile(BufferedReader):
             if offset > 0:
                 raise ValueError('Use negative offsets when seeking backward from end of file')
 
-            # Do we need to generate the total number of datagrams w/in the file?
+            #Do we need to generate the total number of datagrams w/in the file?
             try:
                 self._set_total_dgram_count()
-                # Throws a value error if _total_dgram_count has alread been set.  We can ignore it
+                #Throws a value error if _total_dgram_count has alread been set.  We can ignore it
             except ValueError:
                 pass
 
@@ -550,8 +647,7 @@ class RawSimradFile(BufferedReader):
         elif whence == SEEK_CUR:
             pass
         else:
-            raise ValueError(
-                'Illegal value for \'whence\' (%s), use 0 (beginning), 1 (current), or 2 (end)' % (str(whence)))
+            raise ValueError('Illegal value for \'whence\' (%s), use 0 (beginning), 1 (current), or 2 (end)' % (str(whence)))
 
         if offset > 0:
             for k in range(offset):
@@ -559,6 +655,7 @@ class RawSimradFile(BufferedReader):
         elif offset < 0:
             for k in range(-offset):
                 self.skip_back()
+
 
     def reset(self):
         self._current_dgram_offset = 0
